Create experiments container

+++ experiments/run_experiments.sh
@@ -1,18 +1,30 @@
-time=5
-inference=1
-mode="roulette"
+experiment_names=( "old" "new" )
+times=( 5 10 20 )
+inferences=( 0 1 )
+modes=( "roulette" "elitist" "dynamic" )
 
-experiments=( "axios" "commanderjs" )
-experimentFiles=( "axios/lib/**/*.js" "commanderjs/lib/**/*.js" )
+benchmarks=( "axios" "commanderjs" )
+benchmark_files=( "axios/lib/**/*.js" "commanderjs/lib/**/*.js" )
 
-for x in {0..1}
-do
-  for i in {1..10}
-  do
-    docker rm experiment
-    echo "running experiment ${i} for ${experiments[$x]} with files ${experimentFiles[$x]}"
-    docker run --name experiment --env target_root_directory="./benchmark/top10npm/${experiments[$x]}" --env include="./benchmark/top10npm/${experimentFiles[$x]}" syntest/javascript:t${time}-inference${inference}-mode${mode}
-    docker cp experiment:/app/syntest-javascript/syntest "./results/${time}-${inference}-${mode}-${experiments[$x]}-${i}"
+for experiment_name in "${experiment_names[@]}"; do
+  for time in "${times[@]}"; do
+    for inference in "${inferences[@]}"; do
+      for mode in "${modes[@]}"; do
+        if [[ "$inference" == 0 && "$mode" == "elitist" ]]; then
+          break
+        fi
+
+        for x in {0..1}; do
+          for i in {1..10}; do
+            echo "running experiment ex=${experiment_name}time=${time},inference${inference},mode=${mode} trial ${i} for ${benchmarks[$x]} with files ${benchmark_files[$x]}"
+            docker rm experiment
+            docker run --name experiment --env target_root_directory="./benchmark/top10npm/${benchmarks[$x]}" --env include="./benchmark/top10npm/${benchmark_files[$x]}" syntest/javascript:${experiment_name}-t${time}-inference${inference}-mode${mode}
+            docker cp experiment:/app/syntest-javascript/syntest "./results/${experiment_name}-${time}-${inference}-${mode}-${benchmarks[$x]}-${i}"
+          done
+        done
+
+      done
+    done
   done
 done
 
@@ -20,6 +32,5 @@ done
 #docker run --name experiment --env target_root_directory="./benchmark/top10npm/commanderjs" --env include="./benchmark/top10npm/commanderjs/lib/**/*.js" syntest/javascript:t5-inference1-moderoulette
 #docker run --name experiment --env target_root_directory="./benchmark/top10npm/commanderjs" --env include="./benchmark/top10npm/commanderjs/lib/**/*.js" syntest/javascript:t5-inference1-moderoulette
 #docker run --name experiment --env target_root_directory="./benchmark/top10npm/lodash" --env include="./benchmark/top10npm/lodash/result.js" syntest/javascript:t5-inference1-moderoulette
-#
 
 #docker cp experiment:/app/syntest-javascript/syntest ./
